bio*: rewrite
